Create new directiory tests and add some test
